+++ quickstart/eabm_package/plot_scatter.py
@@ -1,0 +1,1 @@
+../../eabm/eabm_package/python/plot_scatter.py