+++ quickstart/eabm_package/single_element_compression.py
@@ -1,0 +1,9 @@
+nominal = {
+    'width': 1.0,
+    'height': 1.0,
+    'global_seed': 1.0,
+    'displacement': -0.1
+}
+mesh_convergence = {
+    'global_seed': [1.0, 0.5, 0.25, 0.125],
+}